requirements.txt
numpy==2.4.2
scikit-learn==1.8.0
scipy==1.17.0
Pillow==12.1.0
psutil==7.2.2
matplotlib==3.10.1
